Compiles!  Now need to debug app.
git-svn-id: https://cotprojects.marine.usf.edu/svn/MiniSEAS/Gumstix/Source@104 63cbca5c-2ead-4e1b-9715-33247cc11c71

+++ SEASServer/dataFileManager.c
@@ -7,7 +7,7 @@ void openDataFile(){
     char filePath[64];
 
     if(!currDataFile){
-        sprintf(filePath,"%sDATA%u.csv",DATAFILEPATH,time(NULL));        
+        sprintf(filePath,"%sDATA%u.csv",DATAFILEPATH,(unsigned int)time(NULL));        
         syslog(LOG_DAEMON|LOG_INFO,"Opening data file @: %s.",filePath);
         currDataFile = fopen(filePath,"w+");
         if(!currDataFile){
